pyproject.toml
[project]
name = "customario-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "MatinKhajavi",email = "mat.khajavi@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "livekit (>=1.0.19,<2.0.0)",
    "livekit-agents (>=1.2.18,<2.0.0)",
    "livekit-plugins-noise-cancellation (>=0.2.5,<0.3.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "claude-agent-sdk (>=0.1.6,<0.2.0)",
    "anthropic (>=0.73.0,<0.74.0)",
    "livekit-plugins-silero (>=1.2.18,<2.0.0)",
    "livekit-plugins-turn-detector (>=1.2.18,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
